Split upgrade l10n
