+++ tower-defense/src/l10n/upgrade/upgrade_kind/mod.rs
@@ -1,0 +1,2 @@
+mod english;
+mod korean;
